chore: rename bronze to opengeh_bronze
