+++ .github/actions/compile-proto/action.yml
@@ -1,6 +1,11 @@
 name: Compile Measurements Proto Contracts
 description: Compile protobuf contracts for measurements subsystem
 
+inputs:
+  path:
+    description: Path to the directory containing the proto files
+    required: true
+    default: ./source/bronze/src/bronze/infrastructure/contracts
 runs:
   using: composite
   steps:
@@ -10,7 +15,7 @@ runs:
     - name: Compile descriptor files
       shell: bash
       run: |
-        pushd './source/bronze/src/bronze/infrastructure/contracts'
+        pushd '${{ inputs.path }}'
         mkdir ./assets
         protoc --include_imports --descriptor_set_out=assets/persist_submitted_transaction.binpb PersistSubmittedTransaction.proto
         protoc --include_imports --descriptor_set_out=assets/submitted_transaction_persisted.binpb SubmittedTransactionPersisted.proto
