+++ .github/workflows/ci-orchestrator.yml
@@ -48,10 +48,12 @@ jobs:
         with:
           filters: |
             is_changed:
-              - source/bronze/src/bronze/infrastructure/contracts/**
+              - source/bronze/src/opengeh_bronze/infrastructure/contracts/**
       - name: Compile protobuf descriptor files
         uses: ./.github/actions/compile-proto
         if: ${{ steps.changes.outputs.is_changed == 'true' }}
+        with:
+          path: source/bronze/src/opengeh_bronze/infrastructure/contracts
 
   #
   # Build Package Matrix
