+++ source/bronze/pyproject.toml
@@ -9,7 +9,7 @@ dependencies = [
 ]
 
 [project.scripts]
-migrate = "bronze.entry_points:migrate"
+migrate = "opengeh_bronze.entry_points:migrate"
 
 [dependency-groups]
 dev = [
@@ -28,12 +28,9 @@ build-backend = "hatchling.build"
 [tool.hatch.metadata]
 allow-direct-references = true
 
-[tool.hatch.build.targets.wheel]
-packages = ["src/bronze"]
-
 [tool.pytest.ini_options]
 # -s ensures that print("<text>") is written to output. 
-addopts = "--cov-report=term --cov-report=xml:coverage.xml --cov=bronze -ra -vv -s"
+addopts = "--cov-report=term --cov-report=xml:coverage.xml --cov=opengeh_bronze -ra -vv -s"
 
 [tool.coverage.report]
 fail_under = 0
