--- source/bronze/src/bronze/entry_points.py
@@ -1,5 +1,0 @@
-﻿import bronze.migrations.migrations_runner as migrations_runner
-
-
-def migrate() -> None:
-    migrations_runner.migrate()
