+++ source/bronze/src/opengeh_bronze/domain/constants/column_names/bronze_measurements_column_names.py
