+++ source/bronze/src/opengeh_bronze/entry_points.py
@@ -1,0 +1,5 @@
+﻿import opengeh_bronze.migrations.migrations_runner as migrations_runner
+
+
+def migrate() -> None:
+    migrations_runner.migrate()
