+++ source/bronze/src/opengeh_bronze/infrastructure/contracts/DecimalValue.proto
