+++ source/bronze/src/opengeh_bronze/infrastructure/contracts/PersistSubmittedTransaction.proto
