+++ source/bronze/src/opengeh_bronze/infrastructure/contracts/SubmittedTransactionPersisted.proto
