+++ source/bronze/src/opengeh_bronze/infrastructure/contracts/decimal_value.py
